requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
Django==1.7
django-redis==4.3.0
djangorestframework==3.5.3
kombu==3.0.33
Pillow==2.9.0
psycopg2==2.6.1
py-dateutil==2.2
python-dateutil==2.6.0
pytz==2015.7
redis==2.10.5
six==1.9.0
urwid==1.3.1
